refactoring : extract annotation generator
