+++ sporniket-javabeans-doclet/src/main/java/com/sporniket/libre/javabeans/doclet/basic/BasicJavabeanGenerator.java
@@ -1,12 +1,10 @@
 package com.sporniket.libre.javabeans.doclet.basic;
 
-import java.util.List;
+import static com.sporniket.libre.javabeans.doclet.basic.Utils.NEXT_INDENTATION;
+
 import java.util.function.Consumer;
 
 import com.sporniket.libre.javabeans.doclet.JavabeanGenerator;
-import com.sporniket.libre.javabeans.doclet.codespecs.AnnotationParameterSpecs;
-import com.sporniket.libre.javabeans.doclet.codespecs.AnnotationParameterSpecsSingleValue;
-import com.sporniket.libre.javabeans.doclet.codespecs.AnnotationParameterSpecsValuesArray;
 import com.sporniket.libre.javabeans.doclet.codespecs.AnnotationSpecs;
 import com.sporniket.libre.javabeans.doclet.codespecs.FieldSpecs;
 import com.sporniket.libre.javabeans.doclet.codespecs.ImportSpecs;
@@ -45,40 +43,6 @@ import com.sporniket.libre.lang.string.StringTools;
  */
 public class BasicJavabeanGenerator extends BasicGenerator implements JavabeanGenerator
 {
-	private static final String NEXT_INDENTATION = "    ";
-
-	private String computeAnnotationValue(AnnotationParameterSpecs annotation, String indentation)
-	{
-		if (annotation instanceof AnnotationParameterSpecsSingleValue)
-		{
-			return computeAnnotationValue((AnnotationParameterSpecsSingleValue) annotation, indentation + NEXT_INDENTATION);
-		}
-		else if (annotation instanceof AnnotationParameterSpecsValuesArray)
-		{
-			return computeAnnotationValue((AnnotationParameterSpecsValuesArray) annotation, indentation + NEXT_INDENTATION);
-		}
-		throw new IllegalStateException("Unsupported annotation parameter specs.");
-	}
-
-	private String computeAnnotationValue(AnnotationParameterSpecsSingleValue annotation, String indentation)
-	{
-		return outputAnnotationParameterValue(annotation);
-	}
-
-	private String computeAnnotationValue(AnnotationParameterSpecsValuesArray annotation, String indentation)
-	{
-		StringBuilder _result = new StringBuilder("{");
-		for (AnnotationParameterSpecsSingleValue _value : annotation.getValues())
-		{
-			if (_result.length() > 1)
-			{
-				_result.append(", ");
-			}
-			_result.append("\n      ").append(indentation).append(outputAnnotationParameterValue(_value));
-		}
-		return _result.append("\n").append(indentation).append("}").toString();
-	}
-
 	private void outputAccessor(FieldSpecs field)
 	{
 		// getter
@@ -103,47 +67,6 @@ public class BasicJavabeanGenerator extends BasicGenerator implements JavabeanGe
 	public void outputAccessors()
 	{
 		getClassSpecs().getFields().stream().filter(f -> f.isDirectlyRequired()).forEach(f -> outputAccessor(f));
-	}
-
-	private void outputAnnotation(AnnotationSpecs annotations, String indentation)
-	{
-		final List<AnnotationParameterSpecs> _parameters = annotations.getParameters();
-		if (null != _parameters && !_parameters.isEmpty())
-		{
-			String _format = indentation + "@%s(%s)\n";
-			final AnnotationParameterSpecs _firstParameter = _parameters.get(0);
-			StringBuilder _parametersValueStatement = new StringBuilder();
-			if (1 == _parameters.size() && "value".equals(_firstParameter.getName()))
-			{
-				_parametersValueStatement.append(computeAnnotationValue(_firstParameter, indentation));
-			}
-			else
-			{
-				String _nextIndentation = indentation + NEXT_INDENTATION;
-				// explicite parameters
-				for (AnnotationParameterSpecs _parameter : _parameters)
-				{
-					if (_parametersValueStatement.length() > 0)
-					{
-						_parametersValueStatement.append(", ");
-					}
-					_parametersValueStatement.append("\n").append(_nextIndentation)//
-							.append(_parameter.getName()).append(" = ")
-							.append(computeAnnotationValue(_parameter, _nextIndentation));
-				}
-				_parametersValueStatement.append("\n");
-			}
-			getOut().printf(_format, annotations.getType(), _parametersValueStatement.toString());
-		}
-		else
-		{
-			getOut().printf(indentation + "@%s\n", annotations.getType());
-		}
-	}
-
-	private String outputAnnotationParameterValue(final AnnotationParameterSpecsSingleValue value)
-	{
-		return (value.isString()) ? String.format("\"%s\"", value.getValue()) : value.getValue().toString();
 	}
 
 	@Override
